reorganise files a bit
